deep-agents-langchain/requirements.txt
langchain==1.3.0
deepagents==0.1.11
langgraph==0.2.42
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.2
structlog==24.4.0
tavily-python==0.3.7
typing_extensions==4.12.2

